Updated Volumes Deletion

+++ src/app/api/admin/products/deleteVolume/[volumeId]/route.ts
@@ -2,7 +2,7 @@ import pool from "@/lib/db";
 
 export async function DELETE(req: Request, { params }: { params: { volumeId: string } }) {
 
-  const { volumeId } = params;
+  const { volumeId } = await params;
 
   if (!volumeId) {
     return new Response(JSON.stringify({ message: 'Volume Id is required.' }), {
@@ -11,6 +11,19 @@ export async function DELETE(req: Request, { params }: { params: { volumeId: str
   }
 
   try {
+    const checkQuery = 'SELECT COUNT(*) FROM products_volumes WHERE volume_id = $1';
+    const checkValues = [volumeId];
+
+    const checkResult = await pool.query(checkQuery, checkValues);
+
+    if (parseInt(checkResult.rows[0].count) > 0) {
+      return new Response(
+        JSON.stringify(
+          { message: 'This volume is currently in use by one or more products. Please remove it from all products before proceeding with the deletion.' }),
+          { status: 400 }
+      );
+    }
+
     const query = 'DELETE FROM volumes WHERE volume_id = $1 RETURNING *';
     const values = [volumeId];
